Create master tpls file to facilitate sharing of global template vars across tasks
