+++ lib/dbReplace.js
@@ -2,10 +2,11 @@
 
 var grunt = require('grunt');
 var shell = require('shelljs');
+var tpls  = require('../lib/tpls');
 
 function dbReplace(search,replace,output_file) {
 
-    var cmd = grunt.template.process( "sed -i '' 's#<%= search %>#<%= replace %>#g' <%= path %>", {
+    var cmd = grunt.template.process( tpls.search_replace, {
         data: {
             search: search,
             replace: replace,
